+++ src/app/native/knowledge/knowledge.component.ts
@@ -1,10 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { IonicModule } from '@ionic/angular';
 
 @Component({
   selector: 'app-native-knowledge',
   templateUrl: './knowledge.component.html',
   styleUrls: ['./knowledge.component.scss'],
-  standalone: false,
+  standalone: true,
+  imports: [IonicModule],
 })
 export class KnowledgeComponent implements OnInit {
   constructor() {}